+++ crates/findex/python/tests/findex_test.py
@@ -216,7 +216,7 @@ class TestFindex(unittest.TestCase):
         # Calling Upsert without setting the proper callbacks will raise an Exception
         with self.assertRaises(Exception):
             self.findex_interface.upsert_wrapper(
-                indexed_values_and_keywords, self.msk, self.label
+                self.msk, self.label, indexed_values_and_keywords, {}
             )
 
         # Set upsert callbacks here
@@ -226,9 +226,7 @@ class TestFindex(unittest.TestCase):
             self.findex_backend.insert_chain,
         )
 
-        self.findex_interface.upsert_wrapper(
-            indexed_values_and_keywords, self.msk, self.label
-        )
+        self.findex_interface.upsert_wrapper(self.msk, self.label, indexed_values_and_keywords, {})
         self.assertEqual(len(self.findex_backend.entry_table), 5)
         self.assertEqual(len(self.findex_backend.chain_table), 5)
 
@@ -266,7 +264,7 @@ class TestFindex(unittest.TestCase):
             Location.from_int(k): v for k, v in self.db.items()
         }
         self.findex_interface.upsert_wrapper(
-            indexed_values_and_keywords, self.msk, self.label
+            self.msk, self.label,indexed_values_and_keywords, {}
         )
 
         # Adding custom keywords graph
@@ -277,7 +275,7 @@ class TestFindex(unittest.TestCase):
             Keyword.from_string('Martia'): ['Marti'],
             Keyword.from_string('Martial'): ['Martia'],
         }
-        self.findex_interface.upsert_wrapper(graph, self.msk, self.label)
+        self.findex_interface.upsert_wrapper(self.msk, self.label, graph, {})
 
         self.assertEqual(len(self.findex_backend.entry_table), 9)
         self.assertEqual(len(self.findex_backend.chain_table), 9)
@@ -334,13 +332,13 @@ class TestFindex(unittest.TestCase):
             Location.from_int(k): v for k, v in self.db.items()
         }
         self.findex_interface.upsert_wrapper(
-            indexed_values_and_keywords, self.msk, self.label
+            self.msk, self.label, indexed_values_and_keywords, {}
         )
 
         new_label = Label.random()
         res = self.findex_interface.search_wrapper(['Sheperd'], self.msk, new_label)
         # new_label cannot search before compacting
-        self.assertEqual(len(res), 0)
+        self.assertEqual(len(res['Sheperd']), 0)
 
         # removing 2nd db line
         del self.db[2]
@@ -351,14 +349,14 @@ class TestFindex(unittest.TestCase):
         self.assertEqual(len(res['Sheperd']), 2)
         # but not the previous label
         res = self.findex_interface.search_wrapper(['Sheperd'], self.msk, self.label)
-        self.assertEqual(len(res), 0)
+        self.assertEqual(len(res['Sheperd']), 0)
 
         # and the keywords corresponding to the 2nd line have been removed
         res = self.findex_interface.search_wrapper(
             ['Martial', 'Wilkins'], self.msk, new_label
         )
-        assert 'Martial' not in res
-        assert 'Wilkins' not in res
+        assert len(res['Martial']) == 0
+        assert len(res['Wilkins']) == 0
 
 
 if __name__ == '__main__':